requirements.txt
Flask
numpy
scipy
fpdf2
